requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.6
pymongo==4.3.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
python-decouple==3.8
django-cors-headers==4.1.0
django-filter==23.3
gunicorn==20.1.0 